Split off test and game functions into files
